Move web reader into own file
